added integration with lambda
Added necessary imports to comply with Lambda's deployment package
specifications and updated the code to simply return the list of players that
are currently online. This will need to change to use a persistent data store in
the future to _actually_ work, but is good enough for a work in progress.

Also added a simple test case with a log event containing a single log message
in it. More test cases will be needed in the future.