enable rule on startup to avoid useless execution
Added a step in the heartbeat monitor to disable the cloudwatch event when
cloding the server, and slightly changed the execution flow of the startup
to explicitely enable the rule as part of the process rather than as part of the
function cleanup. The goal is to avoid having to run the heartbeat monitor when
it is obvious that the server is turned off.