Fix the execution role permissions
I was getting 403 errors on the heartbeat function. After adding the logging
included in this commit, I noticed it was due to the ec2:DescribeInstanceStatus
action. Did some research, and turns out that this action cannot be applied to a
specific instance ARN. To fix the issue, I extracted this specific action into a
separate statement, giving permission to decribe any EC2 instance.